perfect all SQL

+++ bolg-1/src/controller/blog.js
@@ -1,19 +1,19 @@
 const {exec} = require('../db/mysql')
 const getList = (author,keyword) => {   //模拟数据
     //先返回假数据，格式是正确的，但是没有匹配author，keyword
-    let sql = `select * from blogs;`
+    let sql = `select * from blogs where 1=1 `
     if(author){
-        sql+=`and author = '${author}'`
+        sql+=`and author='${author}' `
     }
     if(keyword){
-        sql+=`and title like '%${keyword}'`
+        sql+=`and title like '%${keyword}' `
     }
-    // sql+=`order by createtime desc;`
+    sql+=`order by createtime desc;`
     return exec(sql)
 }
 
 const getDetail = (id) =>{
-    const sql =`select * from blogs where id='&{id}'`
+    const sql =`select * from blogs where id='${id}'`
     return exec(sql).then(rows=>{
         return rows[0]
     })
@@ -24,8 +24,8 @@ const newBlog = (blogData = {})=>{
     const content = blogData.content
     const author = blogData.author
     const createTime = Date.now()
-    const sql = `insert into blogs (title,content,createtime,author)values('${title}','${content}','${author}','${createTime}')`
-    return exec(sql).then(insertData=>{
+    const sql = `insert into blogs (title,content,createtime,author) values ('${title}','${content}','${author}','${createTime}')`
+    return exec(sql).then(insertData=>{    //inserId为执行SQL返回的字段
         return{
             id:insertData.insertId
         }
@@ -35,7 +35,7 @@ const newBlog = (blogData = {})=>{
 const updateBlog = (id,blogData={})=>{
     const title = blogData.title;
     const content = blogData.content
-    const sql =`update blogs set title='${title}',content='${content}'`
+    const sql =`update blogs set title='${title}',content='${content}' where id=${id}`
     return exec(sql).then(updateData=>{
         if(updateData.affectedRows>0){
             return true
@@ -47,8 +47,8 @@ const updateBlog = (id,blogData={})=>{
     
 }
 
-const delBlog=(id)=>{
-    const sql =`delete from blogs where id='${id}'`
+const delBlog=(id,author)=>{
+    const sql =`delete from blogs where id='${id}' and author='${author}' `
     return exec(sql).then(deleteData => {
                 if (deleteData.affectedRows > 0) {
                     return true
